feat: US-01 - Cross-crate DTO deduplication
Create api_types module in velesdb-core with canonical request/response
DTOs shared across velesdb-server and tauri-plugin-velesdb.

- Add utoipa as optional dep behind `openapi` feature for ToSchema derives
- Move all server request/response types to core api_types module
- Server types.rs now re-exports from velesdb_core::api_types (zero duplication)
- Tauri-plugin imports shared defaults and SearchResultResponse from core
- Add `openapi` feature forwarding to server and tauri-plugin crates
